+++ src/Raven.Playground/Program.cs
@@ -2,15 +2,13 @@ using Microsoft.AspNetCore.Components.Web;
 using Microsoft.AspNetCore.Components.WebAssembly.Hosting;
 
 using Raven.Playground;
-using Raven.Playground.Services;
+using Raven.Playground.Worker;
 
 var builder = WebAssemblyHostBuilder.CreateDefault(args);
 builder.RootComponents.Add<App>("#app");
 builder.RootComponents.Add<HeadOutlet>("head::after");
 
 builder.Services.AddScoped(sp => new HttpClient { BaseAddress = new Uri(builder.HostEnvironment.BaseAddress) });
-builder.Services.AddSingleton<PlaygroundFrameworkReferences>();
-builder.Services.AddSingleton<PlaygroundLanguageService>();
-builder.Services.AddSingleton<PlaygroundProgramRunner>();
+builder.Services.AddScoped<PlaygroundWorkerClient>();
 
 await builder.Build().RunAsync();